# restore precompilation on `1.8+`