--- a/.JuliaFormatter.toml
+++ b/.JuliaFormatter.toml
@@ -9,4 +9,5 @@ whitespace_in_kwargs = true
 remove_extra_newlines = true
 whitespace_ops_in_indices = true
 short_to_long_function_def = false
+long_to_short_function_def = true
 annotate_untyped_fields_with_any = false