--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -21,15 +21,10 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - '1.6' # LTS
-          - '1.7' # latest stable
-        os:
-          - ubuntu-latest
-          - windows-latest
-          - macos-latest
-        arch:
-          - x64
-          - x86
+          - '1.6' # LTS (lowest supported `julia` version declared in `Project.toml`)
+          - '1.8' # latest stable
+        os: [ubuntu-latest, windows-latest, macos-latest]
+        arch: [x64, x86]
         exclude:
           # Test 32-bit only on Linux
           - os: macos-latest
@@ -41,36 +36,17 @@ jobs:
             os: ubuntu-latest
 
     steps:
-
-      # Setup environment
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
       - uses: julia-actions/setup-julia@latest
         with:
           version: ${{ matrix.version }}
-      - name: Cache artifacts
-        uses: actions/cache@v1
-        env:
-          cache-name: cache-artifacts
+      - uses: julia-actions/cache@v1
+      - uses: julia-actions/julia-buildpkg@latest
+      - uses: julia-actions/julia-runtest@latest
+      - uses: julia-actions/julia-processcoverage@latest
+      - uses: codecov/codecov-action@v3
         with:
-          path: ~/.julia/artifacts
-          key: ${{ runner.os }}-test-${{ env.cache-name }}-${{ hashFiles('**/Project.toml') }}
-          restore-keys: |
-            ${{ runner.os }}-test-${{ env.cache-name }}-
-            ${{ runner.os }}-test-
-            ${{ runner.os }}-
-
-      # Julia Dependencies
-      - name: Install Julia dependencies
-        uses: julia-actions/julia-buildpkg@latest
-
-      # Run tests
-      - name: Run Graphical test
-        run:  julia --project -e 'using Pkg; Pkg.test(coverage=true)'
-
-      # - name: Codecov
-      #   uses: julia-actions/julia-uploadcodecov@latest
-      #   env:
-      #     CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
+          file: lcov.info
 
   Skip:
     if: "contains(github.event.head_commit.message, '[skip ci]')"