--- a/src/colors.jl
+++ b/src/colors.jl
@@ -3,13 +3,12 @@ invisible() = RGBA{Float64}(0.0, 0.0, 0.0, 0.0)
 
 # the one-arg cases, meant for single colors
 plot_color(s::AbstractString) = parse(RGBA{Float64}, s)
-function plot_color(s::Symbol)
+plot_color(s::Symbol) =
     try
         parse(RGBA{Float64}, s)
     catch e
         is_colorscheme(s) ? cgrad(s) : rethrow(e)
     end
-end
 plot_color(b::Bool) = b ? error("plot_color(true) not allowed.") : invisible()
 plot_color(::Nothing) = invisible()
 plot_color(c::Colorant) = convert(RGBA{Float64}, c)