--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -267,6 +267,6 @@ end
 
 @testset "allocations" begin  # see PlotUtils.jl/pull/136
     stats = @timed optimize_ticks(0.1123, 100.132)
-    @test stats.bytes < 1_000  # ~ 848 (on 1.6)
-    @test stats.time < 1e-3  # ~ 0.45ms (on 1.6)
+    @test stats.bytes < 1_000  # ~ 736 (on 1.8)
+    @test stats.time < 1e-3  # ~ 0.56ms (on 1.8)
 end