Use `SnoopPrecompile` to restore precompile statements on `1.8+` - no allocation regression as in github.com/JuliaPlots/PlotUtils.jl/pull/136 (where we had to disable precompilation) .

**master**
```julia
@time_imports using Dates, PlotUtils
[...]
    268.6 ms  PlotUtils
@time optimize_ticks(.1123, 100.132);
  0.000539 seconds (6 allocations: 736 bytes)
@time optimize_datetime_ticks(Dates.value(DateTime(2_000)), Dates.value(DateTime(2_100)));
  0.481606 seconds (118.30 k allocations: 5.540 MiB, 27.34% gc time, 99.96% compilation time)
@time cgrad([:red, :blue], Float64[0, 1]);
  0.698911 seconds (626.29 k allocations: 33.006 MiB, 6.75% gc time, 99.91% compilation time)
```

**pr**
```julia
@time_imports using Dates, PlotUtils
[...]
    411.4 ms  PlotUtils
@time optimize_ticks(.1123, 100.132);
  0.000550 seconds (6 allocations: 736 bytes)
@time optimize_datetime_ticks(Dates.value(DateTime(2_000)), Dates.value(DateTime(2_100)));
  0.316230 seconds (15.31 k allocations: 692.859 KiB, 99.93% compilation time)
@time cgrad([:red, :blue], Float64[0, 1]);
  0.479295 seconds (18.01 k allocations: 832.094 KiB, 99.80% compilation time)
```

The load time increase (~+150ms) is easily compensated by compilation time saves.